requirements.txt
django==2.2.2
django-phonenumber-field==3.0.1
phonenumberslite==8.10.14
